+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Launch Args to selectively install hooks in Settings, Data and per-mod in Control stage. Settings and Data now default off and must be enabled if desired, Control defaults to hook all mods.
 * Launch Arg factorioPath is no longer required - if absent or empty, will prompt for location
 * More correct handling of tables with unusual numeric keys
+* Follow `terminal.integrated.shell.*` and `terminal.integrated.automationShell.*` settings
 
 ## 0.18.15